+++ package.json
@@ -8,7 +8,7 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
-    "bootstrap": "^5.2.1",
+    "bootstrap": "^5.2.0",
     "bootstrap-icons": "^1.9.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
