+++ src/components/SearchInput.js
@@ -1,23 +1,26 @@
-import React, { useState } from "react";
+import React from "react";
+import { Form } from "react-bootstrap";
+import { useDispatch } from "react-redux";
+import { search } from "../features/countries/countriesSlice";
 
-const SearchInput = ({ onSearch }) => {
-  const [input, setInput] = useState("");
-
-  const handleSubmit = e => {
-    e.preventDefault();
-
-    onSearch(input);
-  };
+const SearchInput = () => {
+  const dispatch = useDispatch();
 
   return (
-    <form onSubmit={handleSubmit}>
-      <input
-        type="text"
-        placeholder="Find a country"
-        value={input}
-        onChange={e => setInput(e.target.value)}
-      />
-    </form>
+    <div>
+      <Form>
+        <Form.Control
+          style={{ width: "18rem" }}
+          type="search"
+          className="me-2 "
+          placeholder="Search for countries"
+          aria-label="Search"
+          // onChange={(e) => setSearch(e.target.value)}
+          onChange={(e) => dispatch(search(e.target.value))}
+        />
+        <p>clear search</p>
+      </Form>
+    </div>
   );
 };
